refactor: move processing related component to separate package
[SW-588]
